use link0 to use llprio in transmitted packets, regardless of mbuf prio.
some ISPs now provide services over vlans, but require vlan packets
going to the internet have their priority set to 1 (ie, 0 on the
wire) or they will be dropped. configuring this on openbsd requires
config in several places, eg, pf rules to set the prio on ip packets,
llprio on the pppoe interface for it's management frames, and the
llprio on the vlan interface if you're using dhclient on it. this
has the side effect that you can no longer use priority queuing,
and can be error prone to get right.

using link0 to flatteng the priority for isp transit is simple to
configure, and allows priority queuing.

a man page update is coming.

ok henning@