add support for rfc 6391: flow-aware transport of pseudowires.
this basically adds a dummy mpls tag to the stack for pseudowires
and uses a flow as the label on that dummy tag. this allows
intermediate systems that hash packets onto multiple links to use
the extra tag as input to the hash, providing more entropy and
therefore better load balancing.

it's a pity there's no way to turn it on yet...