MC146818_GETTOD: Check for RTC rollover during read.
Because the RTC is still ticking it is theoretically possible for the
second to roll over while we are reading the clock.  This is basically
impossible on real/practical hardware but is an interesting corner case
for e.g. a VM reading an emulated MC146818 during an exit to userspace.

This doesn't check *every* register, so if it takes you a multiple of
60 seconds to read the RTC this change won't help you.  But you've got
bigger problems at that point.

"no objections" kettenis@/mlarkin@, "fine idea" deraadt