support hardware timestamping for received packets.
the chip has a cycle counter or something, and the value of the
counter when the packet is sent to the host is stored in the
completion queue entry.

the driver periodically checks the cycle counter and records the
time at that point, so the rx path can do some maths to figure out
what the clock time is for the cycle counter. if the driver is
calibrated with the chips counter, the rx packets are timestamped.

this is based on the changes made to freebsd, but with some
simplifications.