tweak mtu handling and propagate mtu setting to trunkports
make setting a trunkports mtu to its current mtu a nop. set a
trunkports mtu to the aggr mtu when the port is getting added. set
the mtu on all trunkports when the aggr mtu is set so things look
consistent. restore a trunkports mtu when it is removed from an
aggr.

this is mostly cosmetic since the mtu on trunkports isn't really
used anywhere.