tweak config pending handling so it waits for all ports to finish scanning
it waited until any port finished scanning before calling
config_pending_decr, but if you have disks on other ports then they
may race with the rest of the kernel getting to mountroot. this
tweaks the config pending dance so it waits for all ports and avoids
the race.

i just got lucky on my system where root was on the first port.
kettenis has been poking a t4 where the disks are all over the place
and he lost the race.

tested by and ok kettenis@