default (ie, reset) the partner info when a ports link goes down.
this doesnt seem to be mentioned in the spec, but is a sensible
thing to do if you think about it. all the switches i've tried also
do this, so there's some consensus about it being sensible.

this is done in the link state handler rather than being added to
one of the state machines. the idea is to keep the state machines
as close to what's in the spec as possible.