per-process itimers: itimerval -> itimerspec
Loongson runs at 128hz.  128 doesn't divide evenly into a million,
but it does divide evenly into a billion.  So if we do the per-process
itimer bookkeeping with itimerspec structs we can have error-free
virtual itimers on loongson just as we do on most other platforms.

This change doesn't fix the virtual itimer error alpha, as 1024 does not
divide evenly into a billion.  But this doesn't make the situation any
worse, either.

ok deraadt@