ciss(4): ciss_cmd: split polling loop into two loops
There are really two loops here.  One for SCSI_NOSLEEP and delay(9),
the other for tsleep(9).  The paths share a loop in the code but they
shouldn't.  They have different termination conditions and share little
code.

If we pull the conditional clause out of the for-loop and then give each
path its own polling loop the logic should be easier to follow.

ok krw@