Make acpivout(4) call ws_[gs]et_param instead of directly calling the ACPI methods.  On some machines, like my X395, those ACPI methods don't allow changing the brightness, so this allows acpivout(4) to e.g. use amdgpu(4)'s code.
In comparison to the previously committed and backed out
version we now schedule the brightness change on the ACPI
task queue so that we don't run into any recursive locking
issues.  Additionally the diff has been modified according
to the recent commit where we make sure that a 5% brightness
change does not result in the same brightness level.

ok jcs@ kn@