Add support for the RK3399's VOP clocks to rkclock(4).
The main VOP DCLK clocks either have a normal divisor as a parent
or some fractional one.  The clock referred in the device tree to
set the frequency on is the mux itself, which has no divisor.  To
To be able to set the frequency, we need to allow setting it on the
parent anyway.  Thus move the divisor mask check to a later point,
so even though we have no divisor, we can still tell the parent to
adjust itself.

ok kettenis@