Make it possible to use ix(4) with MSI-X, currently disabled by default.
The current implementation still uses a single queue but already establishes
a different handler for link interrupts.  This is done in preparation for
multi-queues support.

The performance and i386 regression exposed by the last version of this
diff have been fixed thanks to the help of Hrvoje Popovski.

Based on a bigger diff from haesbaert@ and on the FreeBSD code.

Tested by Sigi Rudzio, Hrvoje Popovski and jmatthew@, ok jmatthew@