Rewrite m88k mutex code as a slight variation of the MI mutex code.
This will make mutex spinning time visible in top(1), and also might
improve stability.

The major change in this is that the old assembly code acquires
mutexes with an atomic exchange operation, but releases them with a
regular store, but the new code always uses atomic exchange
operations.

The mutex.h changes to the macros conform to <sys/mutex.h> to be able
to reset the system while in ddb.

Suggested from Miod Vallat, tested by me.  The stability in heavy load
is greatly improved in my case.