raise max columns and rows in efifb to 160
This is the same change made in rev 1.21 to match the drm drivers.
It was reverted as Lucas Raab reported problems with inteldrm taking
over the fb with a 4k display.  Lucas confirmed that this is no longer
an issue.

Prompted by a similar patch from John Carmack to raise the limits.

ok kettenis@